+++ src/Fable.Build/Quicktest/TypeScript.fs
@@ -14,7 +14,8 @@ let handle (args: string list) =
 
     // Make sure the output directory exists, so nodemon doesn't complain
     Directory.CreateDirectory(outDir) |> ignore
-    FileInfo(mainFile).Create() |> ignore
+    // Make sure to dispose the file handle, to avoid file locking
+    FileInfo(mainFile).Create() |> _.Dispose()
 
     let tscCommand =
         CmdLine.empty