+++ src/fable-metadata/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2.0.0-beta.1 - 2025-02-16
+
+* F# 9 support
+
 ## 1.1.0 - 2024-05-23
 
 ### Fixed